Updated Readme.md. Update test classes. In this commit: 1- The `Readme.md` has been updated with the project description 2- The use case has been updated to return the domain object.
