+++ app/src/main/java/com/example/otchallenge/api/dto/BookResultDto.kt
@@ -5,4 +5,8 @@ import com.google.gson.annotations.SerializedName
 data class BookResultDto(
     @SerializedName("display_name") var listName: String,
     var books: List<BookDto>
-)
+) {
+    companion object {
+        val DEFAULT = BookResultDto("", listOf(BookDto.DEFAULT))
+    }
+}
