+++ app/src/main/java/com/example/otchallenge/booklist/BookListPresenter.kt
@@ -1,7 +1,11 @@
 package com.example.otchallenge.booklist
 
+import com.example.otchallenge.api.dto.BookDto
+import com.example.otchallenge.api.dto.BookResponseDto
+import com.example.otchallenge.booklist.uistate.BookUIState
 import com.example.otchallenge.booklist.uistate.BooksContentUIState
 import com.example.otchallenge.booklist.uistate.LoadingState
+import com.example.otchallenge.utils.capitalizeWords
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
@@ -27,7 +31,7 @@ class BookListPresenter @Inject constructor(val view: BookListView, val useCase:
         view.ownerLifecycleScope.launch {
             val state = try {
                 // ** Load the books **
-                val freshContent = useCase()
+                val freshContent = map(useCase())
 
                 // ** Update the UI state, by adding the new books **
                 freshContent.copy(books = uiState.books + freshContent.books, loadingState = LoadingState.Ready)
@@ -51,5 +55,28 @@ class BookListPresenter @Inject constructor(val view: BookListView, val useCase:
         }
 
         loadBooks()
+    }
+
+
+    /**
+     * Reduces the Dto to the UI state
+     * */
+    private fun map(response: BookResponseDto): BooksContentUIState {
+        return BooksContentUIState.EMPTY.copy(
+            books = response.result.books.map { map(it) },
+            listName = response.result.listName,
+            lastModified = response.lastModified
+        )
+    }
+
+    private fun map(bookDto: BookDto): BookUIState {
+        return BookUIState(
+            //** Making the title capitalized (Word by word)*/
+            title = bookDto.title.capitalizeWords(),
+            author = bookDto.contributor,
+            description = bookDto.description,
+            imageUrl = bookDto.bookImage,
+            rank = bookDto.rank
+        )
     }
 }
